requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization & Explainability
matplotlib>=3.7.0
seaborn>=0.12.0
grad-cam>=1.4.0

# Web Framework & Deployment
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
wandb>=0.15.0

# Video Processing
ffmpeg-python>=0.2.0

# Protobuf version constraint (to avoid conflicts with mysql-connector-python if installed)
# If you don't need mysql-connector-python, uninstall it: pip uninstall mysql-connector-python
# Otherwise, this version should work with most packages
protobuf>=4.21.0,<4.26.0

# Optional: For better performance
# Install CUDA-enabled PyTorch if GPU available
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

